requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
python-dateutil>=2.8.2
pandas>=2.1.0
numpy>=1.24.0
scapy>=2.5.0
yara-python>=4.3.1
elasticsearch>=8.10.0
pymisp>=2.4.175
virustotal-python>=1.0.0
aiohttp>=3.8.5
python-magic>=0.4.27
psutil>=5.9.5
py7zr>=0.20.5
python-evtx>=0.7.4
beautifulsoup4>=4.12.2
lxml>=4.9.3
plotly>=5.16.1
dash>=2.13.0
reportlab>=4.0.4
ipaddress>=1.0.23 